Lift all query param-related behavior to widget page
