+++ components/Widget.tsx
@@ -1,49 +1,34 @@
-import { NextRouter, useRouter } from 'next/router';
 import { useCallback, useState } from 'react';
 import Giscus from '../components/Giscus';
 import { AuthContext, getLoginUrl } from '../lib/context';
-import { useIsMounted } from '../lib/hooks';
 import { createDiscussion } from '../services/giscus/createDiscussion';
 import { getToken } from '../services/giscus/token';
 
 interface IWidgetProps {
+  origin: string;
+  session: string;
   repo: string;
-  term?: string;
-  number?: number;
+  term: string;
+  number: number;
   repoId: string;
   categoryId: string;
-  description?: string;
-  reactionsEnabled?: boolean;
-}
-
-function getSession(router: NextRouter) {
-  const session = router.query.session as string;
-  if (session) {
-    const query = { ...router.query };
-    delete query.session;
-    const url = { pathname: router.pathname, query };
-    const options = { scroll: false, shallow: true };
-    router.replace(url, undefined, options);
-  }
-  return session || '';
+  description: string;
+  reactionsEnabled: boolean;
 }
 
 export default function Widget({
+  origin,
+  session,
   repo,
   term,
   number,
   repoId,
   categoryId,
   description,
-  reactionsEnabled = true,
+  reactionsEnabled,
 }: IWidgetProps) {
-  const router = useRouter();
-  const isMounted = useIsMounted();
   const [token, setToken] = useState('');
   const [isFetchingToken, setIsFetchingToken] = useState(false);
-
-  const session = getSession(router);
-  const origin = (router.query.origin as string) || (isMounted ? location.href : '');
 
   const handleDiscussionCreateRequest = async () =>
     createDiscussion(repo, {
@@ -70,8 +55,7 @@ export default function Widget({
       .catch((err) => handleError(err?.message));
   }
 
-  const ready =
-    router.isReady && (!session || token) && !isFetchingToken && repo && (term || number);
+  const ready = (!session || token) && !isFetchingToken && repo && (term || number);
 
   return ready ? (
     <AuthContext.Provider value={{ token, origin, getLoginUrl }}>
